Optimisation des requêtes

- Deux requêtes en moins pour les commentaires des contenus
- N requêtes en moins pour les commentaires des contenus et les messages des forums

**Avant**

- Article avec deux pages de commentaires : 96 requêtes (2s) connecté et 29 requêtes (0,8s) anonyme
- Sujet avec deux pages de messages : 113 requêtes (2,3s) connecté et 18 requêtes (0,7s) anonyme

**Après**

- Article avec deux pages de commentaires : 69 requêtes (1,5s) connecté et 26 requêtes (0,8s) anonyme
- Sujet avec deux pages de messages : 93 requêtes (1,9s) connecté et 18 requêtes (0,7s) anonyme

**QA :**
- Vérifier que les pings fonctionnent correctement dans les commentaires des contenus et les messages du forum.