@artragis 2021-04-25
Pour ma part c'est ok puisque j'avais testé moi-même avant d'ajouter quelques optimisations. Il faudrait donc qu'une revue supplémentaire arrive pour qu'on valide ça.