@philippemilink 2021-04-27
C'est important les *avec deux pages de messages/commentaires* ? Ou c'est juste pour informer que quand tu as relevé les stats, il y avait deux pages ?