@Situphen 2021-04-27
> C'est important les _avec deux pages de messages/commentaires_ ? Ou c'est juste pour informer que quand tu as relevé les stats, il y avait deux pages ?

Ce n'est pas important en tant que tel de mémoire, c'est simplement que avant pour certaines choses il y avait N requêtes pour N commentaires et que maintenant c'est une seule requête pour N commentaires donc la baisse se voit mieux quand il y a pas mal de commentaires. :)