@Situphen 2021-04-27
> La pluralisation du message de recherche d'aide sur un contenu en bêta (L’auteur de ce contenu recherche un illustrateur. N’hésitez pas à le contacter par MP pour proposer votre aide !) ne fonctionne pas. C'est peut-être un bug qui n'est pas lié à cette PR (j'essaie de le reproduire sur la branche dev plus tard).

J'ai vérifié et c'est bel et bien un bug causé par la PR. J'ai donc décidé de ne laisser que les deux premiers commits QA dans cette PR et de déplacer le troisième commit qui demande encore du travail dans une autre PR (#6110). Donc tu peux refaire la QA de celle-ci et la fusionner si c'est bon ! L'autre PR a un gros risque de régression et demande encore du travail apparemment.